minor changes in main class

+++ Group1_F2019_MAD3004_FP/main.swift
@@ -60,7 +60,7 @@ func readJsonFileArray(jsonFileName: String)
                     //print(age)
                     user.age = age
                 }
-                if let hrs = jsonDictionay["hoursworked"] as? Int
+                if let hrs = jsonDictionay["hoursworked"] as? Float
                                {
                                    //print(age)
                                    user.hourswork = hrs
